excele_reporter/requirements.txt
flask==2.3.2
fdb==2.0.2
pandas==2.0.3
openpyxl==3.1.2
XlsxWriter==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
wheel==0.41.1
pymongo==4.4.1